+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.9.1 (July 1, 2018)
+
+IMPROVEMENTS
+
+- RangeProof.ComputeRootHash() to compute root rather than provide as in Verify(hash)
+- RangeProof.Verify\*() first require .Verify(root), which memoizes
+
 ## 0.9.0 (July 1, 2018)
 
 BREAKING CHANGES